requirements.txt
###Algorithms

#Testing
pylint >= 2.14.2

#Google
google-cloud >= 0.34.0
google-cloud-storage >= 1.38.0
google-cloud-bigquery >= 2.10.0
gspread >= 4.0.1
gspread-dataframe >= 3.2.1

#Slack
slack_sdk >= 3.0.0, < 4.0.0

#PDF
pdfminer >= 20191125
weasyprint == 52.5

#Web
requests >= 2.23.0
selenium >= 4.0.0
undetected-chromedriver >= 3.0.3
selenium-wire >= 4.5.4
xvfbwrapper >= 0.2.9 

#Other
numpy >= 1.22.4
pandas >= 1.4.2
tabula-py == 2.2.0
pytz >= 2021.1
lxml >= 4.9.1
cryptography >= 37.0.2

